+++ tests/support/request.js
@@ -1,0 +1,5 @@
+const supertest = require("supertest");
+const app = require('../../src/server.js');
+const request = supertest(app);
+
+module.exports = request;
